requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
